inference_service/requirements.txt
grpcio
onnxruntime-gpu
numpy
opencv-python-headless
prometheus-client